+++ src/menus/mod.rs
@@ -1,0 +1,3 @@
+pub mod credits;
+pub mod pause;
+pub mod start_menu;